Added prop_dynamic_special
Has all features of:
-prop_dynamic
-prop_glow
-prop_scalable

+++ src/game/server/props.cpp
@@ -96,7 +96,7 @@ ConVar sv_turbophysics( "sv_turbophysics", "0", FCVAR_REPLICATED, "Turns on turb
 float GetBreakableDamage( const CTakeDamageInfo &inputInfo, IBreakableWithPropData *pProp )
 {
 	float flDamage = inputInfo.GetDamage();
-	int iDmgType = inputInfo.GetDamageType();
+	int64 iDmgType = inputInfo.GetDamageType();
 
 	// Bullet damage?
 	if ( iDmgType & DMG_BULLET )
